move `focusPath` logic out of addon into hover providers entrypoint
